+++ dio/CHANGELOG.md
@@ -7,6 +7,7 @@ See the [Migration Guide][] for the complete breaking changes list.**
 
 - Provide fix suggestions for `dart fix`.
 - Fix `receiveTimeout` for streamed responses.
+- Fix cancellation and missing progress handling for streamed responses and downloads when using `IOHttpClientAdapter`.
 
 ## 5.4.0